AI-Agentic-AI-Data-Cleaner-v2.0/requirements.txt
# Core Data Handling
pandas==2.1.3
numpy==1.26.2

# API & Web Interface
fastapi==0.109.2
uvicorn==0.27.0
streamlit==1.30.0
requests==2.31.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
openpyxl==3.1.2  # For reading Excel files

# AI & LangGraph
openai==1.6.1
langchain==0.1.10
langgraph==0.0.28
python-dotenv==1.0.0

# Optional (for validation or extension)
pydantic==2.5.3
